backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.0
asyncpg>=0.30.0
sqlalchemy[asyncio]>=2.0.0
PyJWT[cryptography]>=2.9.0
openai>=1.50.0
pgvector>=0.3.0
python-multipart>=0.0.12
httpx>=0.27.0
pdfplumber>=0.11.0
python-docx>=1.1.0
tiktoken>=0.7.0
qrcode[pil]>=8.0
arq>=0.26.0
redis>=5.1.0
sentry-sdk[fastapi]>=2.0.0
